+++ reusables/dt.py
@@ -98,6 +98,12 @@ class DateTime(datetime.datetime):
             desired_format = exp.sub(strf, desired_format)
         return self.strftime(desired_format.format(*args, **kwargs))
 
+    def __repr__(self):
+        tz = ", tzinfo={}".format(self.tzinfo) if self.tzinfo else ""
+        return ("DateTime(year={0.year}, month={0.month}, day={0.day}, "
+                "hour={0.hour}, minute={0.minute}, second={0.second}, "
+                "microsecond={0.microsecond}{1})".format(self, tz))
+
     def __iter__(self):
         for k, v in self.__dict__.items():
             yield (k, v)
@@ -120,7 +126,61 @@ class DateTime(datetime.datetime):
         except ValueError:
             return cls.strptime(iso_datetime, "%Y-%m-%dT%H:%M:%S")
 
-    # TODO add a 'from datetime'
+    @staticmethod
+    def from_datetime(dt):
+        return DateTime(dt.year, dt.month, dt.day, dt.hour, dt.minute,
+                        dt.second, dt.microsecond, dt.tzinfo)
+
+    # Overwrite parent class's creations
+    @classmethod
+    def now(cls, tz=None):
+        return cls.from_datetime(super(DateTime, cls).now(tz))
+
+    @classmethod
+    def utcnow(cls):
+        return cls.from_datetime(super(DateTime, cls).utcnow())
+
+    @classmethod
+    def combine(cls, date, time):
+        """Construct a datetime from a given date and a given time."""
+        if not isinstance(date, datetime.date):
+            raise TypeError("date argument must be a date instance")
+        if not isinstance(time, datetime.time):
+            raise TypeError("time argument must be a time instance")
+        return cls(date.year, date.month, date.day,
+                   time.hour, time.minute, time.second, time.microsecond,
+                   time.tzinfo)
+
+    @classmethod
+    def _fromtimestamp(cls, t, utc, tz):
+        """Construct a datetime from a POSIX timestamp (like time.time()).
+
+        A timezone info object may be passed in as well.
+        """
+        return cls.from_datetime(super(DateTime,
+                                       cls)._fromtimestamp(t, utc, tz))
+
+    @classmethod
+    def fromtimestamp(cls, t, tz=None):
+        """Construct a datetime from a POSIX timestamp (like time.time()).
+
+        A timezone info object may be passed in as well.
+        """
+        return cls.from_datetime(super(DateTime, cls).fromtimestamp(t, tz))
+
+    @classmethod
+    def utcfromtimestamp(cls, t):
+        """Construct a naive UTC datetime from a POSIX timestamp."""
+        return cls.from_datetime(super(DateTime, cls).utcfromtimestamp(t))
+
+    @classmethod
+    def strptime(cls, date_string, format):
+        """
+        string, format -> new datetime parsed from a string
+        (like time.strptime()).
+        """
+        return cls.from_datetime(super(DateTime,
+                                       cls).strptime(date_string, format))
 
 
 def now(utc=False, tz=None):